+++ src/App.css
@@ -6,8 +6,3 @@
     margin: 0;
     padding: 0;
 }
-
-.logo {
-    height: 50px;
-    margin-left: 30px;
-}
