+++ src/components/Navbar/Navbar.css
@@ -1,0 +1,4 @@
+.Navbar__logo {
+  height: 50px;
+  margin-left: 30px;
+}
